♻ REFACTOR

+++ sun_front/src/Component/Nav/Header.js
@@ -3,7 +3,13 @@ import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import NavDropdown from 'react-bootstrap/NavDropdown';
+import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
+import { getData, getUserData, isAccessTokenExpired, logOut, refreshAccessToken } from '../../apis/api/User';
+import { setToken } from '../../redux/reducers/AuthReducer';
+import { Cookies, useCookies } from 'react-cookie';
+import { message } from 'antd';
+import { setLoginState } from '../../redux/reducers/LoginStateReducer';
 
 const Logo = <img src={process.env.PUBLIC_URL + '/img/Logo.png'} alt="logo" style={{
     height: '35px',
@@ -11,41 +17,42 @@ const Logo = <img src={process.env.PUBLIC_URL + '/img/Logo.png'} alt="logo" styl
     marginLeft: '10px'
 }} />;
 
-function Header() {
+function Header(props) {
     const navigate = useNavigate();
+    const dispatch = useDispatch();
     const [auth, setAuth] = useState('');
     const [name, setName] = useState('');
-    const [role, setRole] = useState('');
+    const [role, setRole] = useState(''); 
+    const loginState = props.loginState
+    // const loginState = useSelector((state) => state.LoginState.login_state);
+    const userInfo = useSelector((state) => state.UserInfo);
+
 
     useEffect(() => {
-        if (localStorage.getItem('token') !== null && localStorage.getItem('token') !== undefined && localStorage.getItem('token') !== 'undefined' && localStorage.getItem('token') !== 'null' && localStorage.getItem('token') !== "") {
-            console.log(localStorage.getItem('token'));
-            setAuth(true);
-            setName(localStorage.getItem('name'));
-            setRole(localStorage.getItem('role'));
-        } else if (localStorage.getItem('token') === null || localStorage.getItem('token') === 'null') {
+        if (!loginState) {
             setAuth(false);
+            dispatch(setLoginState(false));
             navigate('/login');
+        } else {
+            setAuth(true);
+            setName(userInfo.last_name + userInfo.first_name);
+            setRole(userInfo.role);
         }
-    }, [auth])
+    }, [loginState])
+
 
     const handleLogout = () => {
-        fetch('http://127.0.0.1:8000/auth/logout/', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Token ${localStorage.getItem('token')}`
-            }
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
+        logOut().then(response => {
+            if (response === 200) {
                 localStorage.clear();
                 setAuth(false);
                 setName("");
                 setRole("");
-                window.location.replace('http://localhost:3000/login');
-            });
+                dispatch(setLoginState(false));
+                navigate('/login');
+                // window.location.replace('http://localhost:3000/login');
+            }
+        })
     };
 
     return (
@@ -60,9 +67,10 @@ function Header() {
                             auth && (
                                 <>
                                     <Nav className="me-auto">
-                                        <Nav.Link href="/center/list">나의 센터</Nav.Link>
                                         <NavDropdown title="센터 등록" id="basic-nav-dropdown">
-                                            {role === "manager" ? <NavDropdown.Item href="/center/register">센터 새로 등록</NavDropdown.Item> : <NavDropdown.Item href="/center/search">센터 ID로 검색</NavDropdown.Item>}
+                                            {role === "manager" ?
+                                                <NavDropdown.Item href="/center/register">센터 새로 등록</NavDropdown.Item>
+                                                : <NavDropdown.Item href="/center/search">센터 검색</NavDropdown.Item>}
                                         </NavDropdown>
                                         {role !== "manager" && (
                                             <Nav.Link href="/timetable">시간표</Nav.Link>
@@ -74,8 +82,8 @@ function Header() {
                         <Navbar.Collapse className="justify-content-end">
                             {auth ? (<>
                                 <NavDropdown title={name} id="basic-nav-dropdown">
-                                    <NavDropdown.Item disabled="true">{role === "manager" ? (<>관리자</>): role === "trainer" ? (<>트레이너</>) : role === "general" ? (<>일반사용자</>) : <></>}</NavDropdown.Item>
-                                    <NavDropdown.Item href="#action/3.2">프로필 수정</NavDropdown.Item>
+                                    <NavDropdown.Item disabled="true">{role === "manager" ? (<>운영자</>) : role === "instroctor" ? (<>강사</>) : role === "general" ? (<>일반사용자</>) : <></>}</NavDropdown.Item>
+                                    <NavDropdown.Item href="/account/profile">프로필 수정</NavDropdown.Item>
                                     <NavDropdown.Divider />
                                     <NavDropdown.Item onClick={handleLogout}>로그아웃</NavDropdown.Item>
                                 </NavDropdown>
